chore: move factor list to risk expr editor
